requirements.txt
feedparser==5.2.1
pytz==2019.1
python-dateutil==2.8.0
